Nota sessions en ekki model

+++ app/src/main/java/is/hi/hbv601g/hopby/entities/Session.java
@@ -117,7 +117,7 @@ public class Session {
         mSlots = mSlots;
     }
 
-    public long getHobbyId() {
+    public int getHobbyId() {
         return mHobbyId;
     }
